requirements.txt
python-dateutil==2.8.2
pandas==2.2.0
pyarrow==15.0.0
openpyxl==3.1.2